requirements.txt
langchain==0.2.16
langchain-groq==0.1.9
langchain-community==0.2.16
langchain-core==0.2.38
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
pypdf==4.3.1
python-dotenv==1.0.1
tiktoken==0.7.0
sentence-transformers==3.0.1
